+++ packages/cryptoplease/lib/app/screens/authenticated/profile/components/menu_header.dart
@@ -11,11 +11,11 @@ import '../../../../../core/tokens/token.dart';
 import '../../../../../gen/assets.gen.dart';
 import '../../../../../l10n/device_locale.dart';
 import '../../../../../l10n/l10n.dart';
+import '../../../../../routes.dart';
 import '../../../../../ui/app_bar.dart';
 import '../../../../../ui/button.dart';
 import '../../../../../ui/colors.dart';
 import '../../../../../ui/icon_button.dart';
-import '../../../../routes.gr.dart';
 
 class MenuHeader extends StatelessWidget {
   const MenuHeader({
