+++ packages/cryptoplease/lib/app/screens/authenticated/profile/menu_screen.dart
@@ -5,10 +5,10 @@ import 'package:flutter/material.dart';
 import '../../../../core/balances/presentation/refresh_balance_wrapper.dart';
 import '../../../../gen/assets.gen.dart';
 import '../../../../l10n/l10n.dart';
+import '../../../../routes.dart';
 import '../../../../ui/colors.dart';
 import '../../../../ui/navigation_bar/navigation_bar.dart';
 import '../../../../ui/theme.dart';
-import '../../../routes.dart';
 import 'components/menu_button.dart';
 import 'components/menu_header.dart';
 import 'components/menu_section.dart';
