+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart
@@ -11,9 +11,9 @@ import '../../../../features/outgoing_split_key_payments/module.dart';
 import '../../../../features/qr_scanner/module.dart';
 import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
+import '../../../../routes.dart';
 import '../../../../ui/dialogs.dart';
 import '../../../../ui/theme.dart';
-import '../../../routes.dart';
 import 'wallet_main_screen.dart';
 
 class WalletFlowScreen extends StatefulWidget {
