+++ packages/cryptoplease/lib/features/backup_phrase/module.dart
@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nested/nested.dart';
 
-import '../../app/routes.dart';
 import '../../core/accounts/bl/accounts_bloc.dart';
 import '../../di.dart';
+import '../../routes.dart';
 import 'src/bl/puzzle_reminder_bloc.dart';
 
 export 'src/routes.dart';
