+++ packages/cryptoplease/lib/features/incoming_split_key_payments/src/presentation/pending_iskp_listener.dart
@@ -2,10 +2,10 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../app/routes.dart';
 import '../../../../core/dynamic_links_notifier.dart';
 import '../../../../core/split_key_payments.dart';
 import '../../../../di.dart';
+import '../../../../routes.dart';
 import '../bl/pending_iskp_repository.dart';
 
 class PendingISKPListener extends StatefulWidget {
