+++ packages/cryptoplease/lib/features/outgoing_direct_payments/src/presentation/build_context_ext.dart
@@ -5,9 +5,9 @@ import 'package:provider/provider.dart';
 import 'package:solana/solana.dart';
 import 'package:uuid/uuid.dart';
 
-import '../../../../app/routes.gr.dart';
 import '../../../../core/amount.dart';
 import '../../../../core/currency.dart';
+import '../../../../routes.dart';
 import '../bl/bloc.dart';
 
 extension OdpRoutesExt on BuildContext {
