+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/src/presentation/oskp_screen.dart
@@ -5,11 +5,11 @@ import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../../app/routes.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../di.dart';
 import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
+import '../../../../routes.dart';
 import '../../../../ui/button.dart';
 import '../../../../ui/content_padding.dart';
 import '../../../../ui/status_screen.dart';
