+++ packages/cryptoplease/lib/features/payment_request/src/presentation/link_request/flow.dart
@@ -4,12 +4,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
 
-import '../../../../../app/routes.dart';
 import '../../../../../core/accounts/bl/account.dart';
 import '../../../../../core/amount.dart';
 import '../../../../../core/conversion_rates/bl/repository.dart';
 import '../../../../../core/user_preferences.dart';
 import '../../../../../di.dart';
+import '../../../../../routes.dart';
 import '../../../../../ui/dialogs.dart';
 import '../../../../../ui/loader.dart';
 import '../../../module.dart';
