+++ packages/cryptoplease/lib/features/pending_activities/src/components/odp_tile.dart
@@ -1,11 +1,11 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../app/routes.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/presentation/format_date.dart';
 import '../../../../gen/assets.gen.dart';
 import '../../../../l10n/device_locale.dart';
+import '../../../../routes.dart';
 import '../pending_activity.dart';
 import 'styles.dart';
