+++ packages/cryptoplease/lib/routes.dart
@@ -1,10 +1,10 @@
 import 'package:auto_route/auto_route.dart';
 
-import '../features/onboarding/module.dart';
-import '../ui/splash_screen.dart';
-import 'screens/authenticated/routes.dart';
+import 'app/screens/authenticated/routes.dart';
+import 'features/onboarding/module.dart';
+import 'ui/splash_screen.dart';
 
-export 'package:cryptoplease/app/routes.gr.dart';
+export 'routes.gr.dart';
 
 @MaterialAutoRouter(
   replaceInRouteName: 'Screen,Route',
